% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{data_update}
\alias{data_update}
\title{Atualiza os bancos de dados de uma biblioteca de funções}
\usage{
data_update(.from_scratch, ...)
}
\arguments{
\item{.from_scratch}{FALSO por definição}

\item{...}{data}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
