% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db.load.table}
\alias{db.load.table}
\title{load db table}
\usage{
db.load.table(dbConnection, tableToLoad)
}
\arguments{
\item{dbConnection}{The DB Connection}

\item{tableToLoad}{The table to load}
}
\description{
DB Helper function to load table
}
\keyword{database}
