% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db.query}
\alias{db.query}
\title{DB Query helper}
\usage{
db.query(dbConnection, sqlQuery)
}
\arguments{
\item{dbConnection}{The DB Connection}

\item{sqlQuery}{The data to save}
}
\description{
DB Helper function to query data - probably unsafe
}
\keyword{database}
