% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db.save.table}
\alias{db.save.table}
\title{save db table}
\usage{
db.save.table(dbConnection, dataToSave, tableToSaveTo, overwrite = TRUE,
  append = FALSE, ...)
}
\arguments{
\item{dbConnection}{The DB Connection}

\item{dataToSave}{The data to save}

\item{tableToSaveTo}{The table to save}
}
\description{
DB Helper function to save table
}
\keyword{database}
