% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db.statement}
\alias{db.statement}
\title{DB Statement helper}
\usage{
db.statement(dbConnection, sqlQuery)
}
\arguments{
\item{dbConnection}{The DB Connection}

\item{sqlQuery}{The data to save}
}
\description{
DB Helper function to execute sql statement - probably unsafe
}
\keyword{database}
