% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db.update.table}
\alias{db.update.table}
\title{Update db table}
\usage{
db.update.table(dbConnection, dataToUpdate, tableToUpdate)
}
\arguments{
\item{dbConnection}{The DB Connection}

\item{dataToUpdate}{The data to Update}

\item{tableToUpdate}{The table to save}
}
\description{
DB Helper function to Update table
}
\keyword{database}
