% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_select.R
\name{font_select}
\alias{font_select}
\title{Seleciona e importa uma fonte}
\usage{
font_select(
  family = "Times New Roman",
  font_import_all,
  font_import_path,
  font_import_pattern,
  load_fonts_device = "pdf"
)
}
\arguments{
\item{family}{Fonte. "Times New Roman" por definição}

\item{font_import_all}{Importa todas as fontes para o \code{R} (Por definição, TRUE)}

\item{font_import_path}{Veja a função {\code{extrafont::font_import}}}

\item{font_import_pattern}{Veja a função {\code{extrafont::font_import}}}

\item{load_fonts_device}{Veja a função {\code{extrafont::font_import}}}
}
\value{
Retorna uma fonte
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Essa função seleciona e importa uma fonte
}
\details{
\code{font_select} Baseada no pacote {\code{extrafont}}.
}
