% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{getUniqueGeneMat}
\alias{getUniqueGeneMat}
\title{getUniqueGeneMat: get matrix of unique values (rows) per gene}
\usage{
getUniqueGeneMat(m, g, w)
}
\arguments{
\item{m}{matrix of value to simplify}

\item{g}{vector of gene names}

\item{w}{vector of weights, for each redndant gene, the highest weight will be kept}
}
\value{
a matrix m with one value per gene
}
\description{
getUniqueGeneMat: get matrix of unique values (rows) per gene
}
