% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github}
\alias{github}
\alias{github_aut}
\alias{github_files}
\alias{github_script}
\alias{github_read_all}
\title{Funções para acessar arquivos e scrips em repositórios privados do github}
\usage{
github_aut(.user, .repo, .personal_token)

github_files(.authenticate = list(), .github_path, .fun = rio::import)

github_script(.authenticate = list(), .github_path)

github_read_all(
  .authenticate = list(),
  .folder,
  .data,
  .labels,
  .fun,
  .ext,
  .envir = FALSE
)
}
\arguments{
\item{.user}{Nome do usuário}

\item{.repo}{Nome do repositório}

\item{.personal_token}{Sys.getenv("TOKEN"). Crie o ambiente de variáveis (usethis::edit_r_environ()) e adicione o seu token pessoal nele.}

\item{.authenticate}{Uma lista que deve ser preenchida com o nome do usuário, repositório e token, nesta ordem. Use a função \link{github_aut} para isso.}

\item{.github_path}{diretório e nome do arquivo ou script com sua respectiva extensão para leitura ou execução.}

\item{.fun}{Função para ler os arquivos (\code{ rio::impor} por definição)}

\item{.folder}{Diretório dos bancos de dados (Exclusivo para a função \code{github_read_all()}).}

\item{.data}{Bancos de dados que serão lidos (Exclusivo para a função \code{github_read_all()}).}

\item{.labels}{Rótulos dos bancos de dados. Por definição, se \code{missing(.labels) = TRUE}, então \code{.labels = .data} (Exclusivo para a função \code{github_read_all()}).}

\item{.ext}{Extensão dos arquivos (Exclusivo para a função \code{github_read_all()}).}

\item{.envir}{Os bancos de dados serão atribuidos ao ambiente global em formato de lista? Por definição, FALSE.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

As funções \code{github_aut}, \code{github_files}, \code{github_script} e \code{github_read_all} acessam arquivos e scrips em repositórios privados do github.

\strong{github_aut()} faz a autenticação de acesso.

\strong{github_files()} lê arquivos de um repositório privado.

\strong{github_script()} executa scripts de um repositório privado

\strong{github_read_all()} lê e carrega um conjunto de bancos de dados de um repositório privado
}
\note{
Inspiração: \href{https://stackoverflow.com/questions/52614317/sourcing-r-files-in-a-private-github-folder}{Sourcing r files in a private github folder}
}
