% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue_collapse2}
\alias{glue_collapse2}
\title{Transforma variáveis em escalares separados por virgula}
\usage{
glue_collapse2(.data, .variable, .distint = FALSE, .last = " e ")
}
\arguments{
\item{.data}{Banco de dados}

\item{.variable}{variável}

\item{.distint}{Considera apenas valores únicos (FALSO por definição)}

\item{.last}{See glue::glue_collapse() documentation}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Esta função transforma colunas em escalares separados por virgula.
}
