% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_split_named.R
\name{group_split_named}
\alias{group_split_named}
\title{Cria uma lista nomeada de bancos de dados agrupados por uma ou mais variáveis}
\usage{
group_split_named(.data, ...)
}
\arguments{
\item{.data}{Banco de dados}

\item{...}{variáveis para agrupamento}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Esta função cria uma lista nomeada de bancos de dados agrupados por uma ou mais variáveis
}
