% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_manual.R
\name{gtable_manual}
\alias{gtable_manual}
\title{Cria uma gtable}
\usage{
gtable_manual(.data = NULL, l = 1)
}
\arguments{
\item{.data}{Banco de dados}

\item{l}{see gtable documentation}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Essa função cria uma gtable
}
