% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_global.R
\name{keep_global}
\alias{keep_global}
\title{Remove objetos}
\usage{
keep_global(.keep)
}
\arguments{
\item{.keep}{Objetos que serão mantidos}
}
\value{
Limpa o ambiente global ou retorna apenas os objetos desejados
}
\description{
Essa função remove objetos do "global environment"
}
