% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{not.na}
\alias{not.na}
\title{@title Check if a value is not NA}
\usage{
not.na(...)
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function check if a value is not a missing value (NA)
}
