% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{package_update}
\alias{package_update}
\title{Atualiza documentação, constroi e reinstala uma biblioteca de funções localmente}
\usage{
package_update(.update_data = FALSE, .update_readme = FALSE, .check = FALSE)
}
\arguments{
\item{.update_data}{Atualiza os bancos de dados (FALSE por definição)}

\item{.update_readme}{Atualiza o arquivo README (FALSE por definição)}

\item{.check}{devtools::check()}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Esta função atualiza documentação, constroi e reinstala uma biblioteca de funções localmente
}
