% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Extrai a porcentagem de um valor}
\usage{
percent(.x, .total, .decrease)
}
\arguments{
\item{.x}{valor}

\item{.total}{100 por default}

\item{.decrease}{To decrease by 100}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Esta função extrai a porcentagem de um valor
}
