% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{qProjICA}
\alias{qProjICA}
\title{Title Simple ICA projection, default using Puleo et al components}
\usage{
qProjICA(
  newexp,
  ICAgw = qutils:::.puleoICA$gw,
  geneNormType = "sc",
  projNormType = "raw",
  ming = 500
)
}
\arguments{
\item{newexp}{A new expression dataset}

\item{ICAgw}{ICA gene weights (S matrix)}

\item{geneNormType}{Normalization of gene expression matrix (sc: sample scale is default)}

\item{projNormType}{Normalization of components (raw:  is default)}

\item{ming}{Minimum number of overlapping genes}
}
\value{
projected components
}
\description{
Title Simple ICA projection, default using Puleo et al components
}
