% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{quantNorm2ref}
\alias{quantNorm2ref}
\title{Title Quantile normalization}
\usage{
quantNorm2ref(x, ref)
}
\arguments{
\item{x}{new expression dataset to be normalied}

\item{ref}{reference dataset to be used as reference for quantile noralisation}
}
\value{
quntile normalized x
}
\description{
Title Quantile normalization
}
