% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all.R
\name{read_all}
\alias{read_all}
\title{Carrega bancos de dados}
\usage{
read_all(
  path,
  ext,
  .fun = rio::import,
  .labels,
  .envir = FALSE,
  .check = TRUE,
  .exclude
)
}
\arguments{
\item{path}{Caminho até o diretório}

\item{ext}{Extensão das bases}

\item{.fun}{Função de leitura}

\item{.labels}{Rótulos dos bancos}

\item{.envir}{Se FALSE, cria uma lista. Se TRUE, atribui os bancos no ambiente global (FALSE por definição)}

\item{.check}{Retorna um objeto com as informações dos bancos (FALSE por definição)}

\item{.exclude}{Exclui bancos de dados}
}
\value{
Retorna uma lista de banco de dados
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Essa função carrega bases de dados armazenados em um diretório.
}
