% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{show.error}
\alias{show.error}
\title{Show Error message}
\usage{
show.error(msg, shouldStop = TRUE)
}
\arguments{
\item{msg}{The message to be output}

\item{shouldStop}{Whether the message should stop execution}
}
\description{
Error level wraper for outputing message to the user
}
\examples{
show.error("Error to show")
show.error("Message to show", FALSE)
}
\keyword{messaging}
\keyword{user}
