% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{show.info}
\alias{show.info}
\title{Show Info message}
\usage{
show.info(msg, shouldStop = FALSE)
}
\arguments{
\item{msg}{The message to be output}

\item{shouldStop}{Whether the message should stop execution}
}
\description{
Info level wraper for outputing message to the user
}
\examples{
show.warning("Message to show")
show.warning("Info to show", FALSE)
}
\keyword{messaging}
\keyword{user}
