% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{show.message}
\alias{show.message}
\title{Show message helper}
\usage{
show.message(msg, shouldStop = FALSE)
}
\arguments{
\item{msg}{The message to be output}

\item{shouldStop}{Whether the message should stop execution}
}
\description{
Message output helper with option to stop execution
}
\examples{
show.message("Message to show")
show.message("Error to show", TRUE)
}
\keyword{messaging}
\keyword{user}
