% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{show.warning}
\alias{show.warning}
\title{Show Warning message}
\usage{
show.warning(msg, shouldStop = FALSE)
}
\arguments{
\item{msg}{The message to be output}

\item{shouldStop}{Whether the message should stop execution}
}
\description{
Warning level wraper for outputing message to the user
}
\examples{
show.warning("Warning to show")
show.warning("Error to show", FALSE)
}
\keyword{messaging}
\keyword{user}
