% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.set.R
\name{Sys.set}
\alias{Sys.set}
\title{Define a localidade dos processos em R}
\usage{
Sys.set()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Esta função definie a localidade dos processos em R
}
