% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Creates descriptive tables}
\usage{
table(data, .var, ...)
}
\arguments{
\item{data}{data frame}

\item{.var}{variable}

\item{...}{\code{gtsummary::tbl_summary} arguments}
}
\value{
Return a pretty neat summary table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function creates descritive tables
}
\details{
\code{table} fuction takes advantage from {\code{gtsummary}} and {\code{dplyr}} packages.
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{# Using this function with `purrr:map()`, you are able to create tables sequentially.
library(magrittr)
list("hp", "mpg") \%>\%
purrr::map(~Qutils::table(mtcars, .x))
}\if{html}{\out{</div>}}
}

